Save idle cpu and timeout at 10 msec
Ideally instead of 10 msec timeout, this should have been
infinity (that is None), but post that change the messaging
stops working because it is possible that this would
block other actions, like handling messages on the queues.

TODO: find a better solution.
Additionally, anything lower than 10 msec would unnecessarily
CPU, because epoll_wait() will have a time out 1 msec if the
here is set to 0.0, which is bad for idle python nodes.
Note that setting a value of 0.01 or 10 msec would imply that
python node would not be able to send replies faster than 10
to Erlang node, but this is an acceptable delay condering the
idle python node cpu consumption issue mentioned above.