Python 3.7 compatibility (#45)
* Bump Gevent and Greenlet's versions

Both libraries depend on a version of Cython that is not compatible with
Python 3.7; I thus updated these dependencies to their most recent
stable version.

* Rename /async directory to async_support

`async` is a reserved python keyword since the 3.7 version and thus
cannot be used by a library. I renamed the directory and corrected the
imports accordingly.

* Corrected `Atom` import in tests

Tests attempt to import `Atom` from the Pyrlang library. However, it
seems that the class' definition has been moved to the Term library. I
simply changed the import and the tests are now passing fine.

* Actualize the docs related to /async