Update Distribution Protocol to work with OTP-25 nodes
=== OTP-25, ERTS User's Guide
https://www.erlang.org/docs/25/apps/erts/erl_dist_protocol.html#distribution-handshake
13.2  Distribution Handshake
The protocol was introduced in Erlang/OTP R6 and amended in OTP 23.
From OTP 25 support for the older protocol was dropped.
Therefor an OTP 25 node can not connect to nodes older than OTP 23.
===

That's OTP, but in this commit attempt was made to support OTP-25
while preserving connectivity to older nodes than OTP-23.